requirements.txt
streamlit==1.24.0
numpy==1.24.3
opencv-python==4.7.0.72
pillow==9.5.0
tensorflow==2.12.0
scipy==1.10.1
matplotlib==3.7.1